requirements.txt
cffi==1.6.0
Django==1.11.15
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
enum34==1.1.4
idna==2.1
ipaddress==1.0.16
MySQL-python==1.2.5
numpy==1.15.1
pandas==0.21.0
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
PyJWT==1.4.0
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.1
scipy==1.1.0
six==1.10.0
